bill Progress

+++ HMSYSTEM/Models/BillDetail.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int BillId {  get; set; }
         public int? ServiceItemId {  get; set; }
-        public decimal? Qty { get; set; }
+        public int? Qty { get; set; }
         public decimal? Amount { get; set; }
         public decimal? TotalAmount { get; set; }
     }
