+++ HMSYSTEM/ViewModels/BillViewModel.cs
@@ -31,7 +31,7 @@ namespace HMSYSTEM.ViewModels
         public int Id { get; set; }
         public int BillId { get; set; }
         public int? ServiceItemId { get; set; }
-        public decimal? Qty { get; set; }
+        public int? Qty { get; set; }
         public decimal? Amount { get; set; }
         public decimal? TotalAmount { get; set; }
     }
